requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
motor==3.0.0
multidict==6.0.2
py-cord==2.0.0rc1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
yarl==1.7.2
